Move anal.lm32 ##arch
