+++ libr/meson.build
@@ -8,6 +8,7 @@ anal_plugins = [ 'null' ]
 arch_plugins = [ 'null',
   'sh',
   'x86_nz',
+  'lm32',
   'mcs96',
   'ws',
   'arm',
@@ -158,7 +159,6 @@ anal_plugins += [
   'i8080',
   'java',
   'lh5801',
-  'lm32',
   'loongarch_gnu',
   'm68k_cs',
   'm68k_gnu',
